requirement.txt
Django==1.8.5
argparse==1.2.1
beautifulsoup4==4.4.1
django-crispy-forms==1.5.2
selenium==2.48.0
wsgiref==0.1.2
